requirements.txt
aiohttp==3.7.4.post0
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.3.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
billiard==3.6.4.0
bleach==4.1.0
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
Django==3.2
django-qr-code==2.1.0
django-redis==4.12.1
django-wkhtmltopdf==3.3.0
django-yaturbo==1.0.0
idna==2.10
importlib-metadata==4.0.1
lxml==4.6.3
matplotlib==3.4.1
multidict==5.1.0
packaging==21.0
paho-mqtt==1.6.1
Pillow==7.2.0
psycopg2-binary==2.8.6
pycurl==7.43.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-telegram-bot==13.5
pytz==2020.1
redis==3.5.3
requests==2.25.1
short-url==1.2.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tornado==6.1
transliterate==1.10.2
typing-extensions==3.7.4.3
tzlocal==2.1
ufw==0.36
unicode==2.7
Unidecode==1.1.1
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.6.3
zipp==3.4.1